// contracts/typechain-types/BioPassportRegistry.ts
export declare namespace BioPassportRegistry {
  export type CredentialStruct = {
    credentialId: string;
    materialId: string;
    credType: BigNumberish;
    commitmentHash: BytesLike;
    issuer: AddressLike;
    issuerId: string;
    issuedAt: BigNumberish;
    validUntil: BigNumberish;
    artifactCid: string;
    artifactHash: BytesLike;
    revoked: boolean;
  };

  export type CredentialStructOutput = [
    credentialId: string,
    materialId: string,
    credType: bigint,
    commitmentHash: string,
    issuer: string,
    issuerId: string,
    issuedAt: bigint,
    validUntil: bigint,
    artifactCid: string,
    artifactHash: string,
    revoked: boolean
  ] & {
    credentialId: string;
    materialId: string;
    credType: bigint;
    commitmentHash: string;
    issuer: string;
    issuerId: string;
    issuedAt: bigint;
    validUntil: bigint;
    artifactCid: string;
    artifactHash: string;
    revoked: boolean;
  };

  export type MaterialStruct = {
    materialId: string;
    materialType: string;
    metadataHash: BytesLike;
    owner: AddressLike;
    ownerOrg: string;
    status: BigNumberish;
    createdAt: BigNumberish;
    updatedAt: BigNumberish;
  };

  export type MaterialStructOutput = [
    materialId: string,
    materialType: string,
    metadataHash: string,
    owner: string,
    ownerOrg: string,
    status: bigint,
    createdAt: bigint,
    updatedAt: bigint
  ] & {
    materialId: string;
    materialType: string;
    metadataHash: string;
    owner: string;
    ownerOrg: string;
    status: bigint;
    createdAt: bigint;
    updatedAt: bigint;
  };

  export type TransferStruct = {
    transferId: string;
    materialId: string;
    fromAddress: AddressLike;
    fromOrg: string;
    toAddress: AddressLike;
    toOrg: string;
    shipmentHash: BytesLike;
    timestamp: BigNumberish;
    accepted: boolean;
  };

  export type TransferStructOutput = [
    transferId: string,
    materialId: string,
    fromAddress: string,
    fromOrg: string,
    toAddress: string,
    toOrg: string,
    shipmentHash: string,
    timestamp: bigint,
    accepted: boolean
  ] & {
    transferId: string;
    materialId: string;
    fromAddress: string;
    fromOrg: string;
    toAddress: string;
    toOrg: string;
    shipmentHash: string;
    timestamp: bigint;
    accepted: boolean;
  };
}